Restructure tests dir.
This commit:
 - Groups tests by what they test, rather than where they originally
   came from.
 - Adds module docstrings.

Before:
```
./lib.rs
./tracing
./ykbh.rs
./test_helpers.c
./helpers.rs
./guardfailure.rs
./ykcompile.rs
./yktrace.rs
```

After:
```
./lib.rs
./tir.rs
./test_helpers.c
./helpers.rs
./tracing.rs
./codegen
./codegen/mod.rs
./codegen/reg_alloc.rs
./stopgap
./stopgap/mod.rs
./stopgap/interp.rs
./stopgap/guard_fail.rs
```
