+++ tests/src/stopgap/interp.rs
@@ -1,3 +1,5 @@
+//! Artificially testing the stopgap interpreter (without a guard failure).
+
 use ykshim_client::interpret_body;
 
 #[test]
