+++ tests/src/stopgap/mod.rs
@@ -1,0 +1,4 @@
+//! Tests for the stopgap interpreter.
+
+mod guard_fail;
+mod interp;
